+++ types/events/functions.d.ts
@@ -1,0 +1,6 @@
+export function triggerDatepickerEvent(datepicker: any, type: any): void;
+export function goToPrevOrNext(datepicker: any, direction: any): void;
+export function switchView(datepicker: any): void;
+export function clearSelection(datepicker: any): void;
+export function goToOrSelectToday(datepicker: any): void;
+export function unfocus(datepicker: any): void;