+++ types/events/pickerListeners.d.ts
@@ -1,0 +1,6 @@
+export function onClickViewSwitch(datepicker: any): void;
+export function onClickPrevButton(datepicker: any): void;
+export function onClickNextButton(datepicker: any): void;
+export function onClickView(datepicker: any, ev: any): void;
+export function onMousedownPicker(ev: any): void;
+export function onChangeTime(datepicker: any, ev: any): void;