+++ types/i18n/base-locales.d.ts
@@ -1,0 +1,15 @@
+declare namespace _default {
+    namespace en {
+        let days: string[];
+        let daysShort: string[];
+        let daysMin: string[];
+        let months: string[];
+        let monthsShort: string[];
+        let today: string;
+        let clear: string;
+        let hour: string;
+        let minute: string;
+        let titleFormat: string;
+    }
+}
+export default _default;