+++ types/lib/dom.d.ts
@@ -1,0 +1,8 @@
+export function parseHTML(html: any): DocumentFragment;
+export function getParent(el: any): any;
+export function isActiveElement(el: any): boolean;
+export function isVisible(el: any): boolean;
+export function hideElement(el: any): void;
+export function showElement(el: any): void;
+export function emptyChildNodes(el: any): void;
+export function replaceChildNodes(el: any, newChildNodes: any): void;