+++ types/options/shortcutKeys.d.ts
@@ -1,0 +1,1 @@
+export default function createShortcutKeyConfig(options: any): {};